Desenvolvimento

+++ Parking.Management.Dashboard/Backend/src/domain/use-cases/vehicle-note/delete/delete-vehicle-note-usecase.ts
@@ -8,13 +8,21 @@ export class DeleteVehicleNoteUseCase {
   constructor(private vehicleNoteRepository: IVehicleNoteRepository) {}
 
   async execute(id: string): Promise<Result<VehicleNoteDto>> {
+    const vehicleNote = this.vehicleNoteRepository.findById(id);
+
+    if (!vehicleNote) {
+      return new Result<VehicleNoteDto>({
+        message: "Anotação de veiculo não encontrada.",
+      });
+    }
+
     const result = await this.vehicleNoteRepository.delete(id);
 
     const vehicleNoteDto = mapper.map(result, VehicleNote, VehicleNoteDto);
 
-    return new Result<VehicleNoteDto>(
-      vehicleNoteDto,
-      "Universidade deletada com sucesso."
-    );
+    return new Result<VehicleNoteDto>({
+      content: vehicleNoteDto,
+      message: "Anotação de veiculo deletada com sucesso.",
+    });
   }
 }
